Class AceEstimate

A class containing information about a single univariate ACE model.

A generic function for extracting the Details slot of an AceEstimation object.

# S4 method for class 'AceEstimate'
GetDetails(object)

Arguments

object

ACE object

Note

The contents of the Details list depends on the underlying estimation routine. For example, when the ACE model is estimated with a DF analysis, the output is an stats::lm() object, because the stats::lm() function was used (i.e., the basic general linear model). Alternatively, if the user specified the lavaan::lavaan() package should estimate that ACE model, the output is a lavaan::lavaan() object.

Objects from the Class

Objects can be created by calls of the form: new("AceEstimate", aSquared, cSquared, eSquared, caseCount, unity, withinBounds, details, ...)

Author

Will Beasley

Examples

library(NlsyLinks) # Load the package into the current R session.

showClass("AceEstimate")
#> Class "AceEstimate" [package "NlsyLinks"]
#> 
#> Slots:
#>                                                                        
#> Name:      ASquared     CSquared     ESquared    CaseCount        Unity
#> Class:      numeric      numeric      numeric      numeric      logical
#>                                 
#> Name:  WithinBounds      Details
#> Class:      logical         list
est <- CreateAceEstimate(.5, .2, .3, 40)
est
#> [1] "Results of ACE estimation: [show]"
#>  ASquared  CSquared  ESquared CaseCount 
#>       0.5       0.2       0.3      40.0 
print(est)
#> [1] "Results of ACE estimation: [show]"
#>  ASquared  CSquared  ESquared CaseCount 
#>       0.5       0.2       0.3      40.0