Instantiate an AceEstimate-class object.

Creates an instance of the S4 class, AceEstimate instantiating arguments set the parameter values estimated by the ACE model.

CreateAceEstimate(aSquared, cSquared, eSquared, caseCount,
  details=list(), unityTolerance = 1e-11)

Arguments

aSquared
The proportion of variability due to a shared genetic influence (typically represented as a^2, or sometimes h^2).

cSquared
The proportion of variability due to shared common environmental influence.

eSquared
The proportion of variability due to unexplained/residual/error influence.

caseCount
The number of cases used to estimate the model.

details
A list that contains the modeling output and details.

unityTolerance
Specifies how close the the sum of the ACE components should be to one, to be considered properly scaled to one.

Value

An S4 object of AceEstimate-class.

Details

The contents of the details list depends on the underlying estimation routine. For example, when the ACE model is estimated with a DF analysis, the output is an lm object, because the lm function was used (ie, the basical general linear model). Alternatively, if the user specified the lavaan package should estimate that ACE model, the output is a lavaan object.